% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TP_N.R
\name{TP_N}
\alias{TP_N}
\title{A Class of TP-N Distribution objects}
\description{
The \code{mfd} class represents functional data ...
}
\examples{
x <- 1

}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{basis}}{a basismfd object}

\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{basis}}{a basismfd object}

\item{\code{coeff}}{a matrix with nrow=subjects and ncol=total number of basis ...}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TP_N-new}{\code{TP_N$new()}}
\item \href{#method-TP_N-pdf}{\code{TP_N$pdf()}}
\item \href{#method-TP_N-cdf}{\code{TP_N$cdf()}}
\item \href{#method-TP_N-clone}{\code{TP_N$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TP_N-new"></a>}}
\if{latex}{\out{\hypertarget{method-TP_N-new}{}}}
\subsection{Method \code{new()}}{
Constructor for TP_N objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TP_N$new(mean = 0, sd = 1, gamma)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mu}}{a scalar}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TP_N-pdf"></a>}}
\if{latex}{\out{\hypertarget{method-TP_N-pdf}{}}}
\subsection{Method \code{pdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TP_N$pdf(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TP_N-cdf"></a>}}
\if{latex}{\out{\hypertarget{method-TP_N-cdf}{}}}
\subsection{Method \code{cdf()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TP_N$cdf(x)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TP_N-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TP_N-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TP_N$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
